+++ CHANGELOG.md
@@ -28,6 +28,8 @@ All changes to the **Hearts of Modding** extension will be documented in this fi
   - Added specific validation for `map/buildings.txt` to warn about empty lines, which are counted as errors by the HOI4 engine.
   - Added a styling exception for `map/buildings.txt` to suppress the end-of-file newline diagnostic, preventing forced empty lines at the end of the file.
 
+- Added an option to bulk fix all styling issues in a file.
+
 ## [0.5.4] - 2026-05-16
 
 ### Added